1st Commit: Preliminary Stage of the game
This is the start of the game development.

My intention is to create a game like the classical space-invaders with
a look alike graphics and sound effects. The traditional game had a
great impact those days and still has its potential today. Simple games
go well with nowadays young audience, as well with the elder ones, who
will remind them of their childhood games. The game will be simple to
play (no need of complex instructions on how to play the game) just few
keys and the game play will get challenging as the player go up in
levels ending up with a boss. At the end of the game a score board will
be shown to see their current ranking. Before playing this game the
score board will be shown as well to urge the need to be the first in
rank.